08_advanced_pdf_parser/requirements.txt
# PDF 智能解析器依赖

# 核心依赖
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0

# PDF 解析增强
unstructured>=0.10.0
pdfplumber>=0.10.0
camelot-py[cv]>=0.11.0
PyMuPDF>=1.23.0

# OCR 识别
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 工具库
tqdm>=4.66.0